backend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
anyio==4.4.0
appdirs==1.4.4
attrs==24.2.0
backoff==2.2.1
boilerpy3==1.0.7
cattrs==24.1.1
certifi==2024.8.30
charset-normalizer==3.3.2
datasets==3.0.0
dill==0.3.8
docopt==0.6.2
Events==0.5
farm-haystack==1.26.3
filelock==3.16.0
frozenlist==1.4.1
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.7
idna==3.10
inflect==7.4.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lazy-imports==0.3.1
MarkupSafe==2.1.5
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.3
num2words==0.5.13
numpy==2.1.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.3.3
posthog==3.6.6
prompthub-py==4.0.0
pyarrow==17.0.0
pydantic==1.10.18
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
quantulum3==0.9.2
rank-bm25==0.2.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-cache==0.9.8
rpds-py==0.20.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
sseclient-py==1.8.0
sympy==1.13.2
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.15.2
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.39.3
triton==3.0.0
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.1
url-normalize==1.4.3
urllib3==2.2.3
xxhash==3.5.0
yarl==1.11.1
